+++ CHANGELOG.md
@@ -21,6 +21,12 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### New
+
+* Added `addExtension()`
+  - based on the older `addExtensions()` function
+  - name better reflects that we can only add one type of extension at a time
+
 ### Deprecated
 
 * `addExtensions()` is now deprecated